+++ tests/AutoMapperBaseTest.php
@@ -9,11 +9,13 @@ use AutoMapper\Extractor\ClassMethodToCallbackExtractor;
 use AutoMapper\Generator\Generator;
 use AutoMapper\Loader\ClassLoaderInterface;
 use AutoMapper\Loader\FileLoader;
+use Doctrine\Common\Annotations\AnnotationReader;
 use PhpParser\ParserFactory;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Serializer\Mapping\ClassDiscriminatorFromClassMetadata;
 use Symfony\Component\Serializer\Mapping\Factory\ClassMetadataFactory;
+use Symfony\Component\Serializer\Mapping\Loader\AnnotationLoader;
 use Symfony\Component\Serializer\Mapping\Loader\AttributeLoader;
 
 /**
@@ -34,11 +36,15 @@ abstract class AutoMapperBaseTest extends TestCase
     {
         $fs = new Filesystem();
         $fs->remove(__DIR__ . '/cache/');
-        $classMetadataFactory = new ClassMetadataFactory(new AttributeLoader());
+        if (class_exists(AttributeLoader::class)) {
+            $classMetadataFactory = new ClassMetadataFactory(new AttributeLoader());
+        } else {
+            $classMetadataFactory = new ClassMetadataFactory(new AnnotationLoader(new AnnotationReader()));
+        }
 
         $this->loader = new FileLoader(new Generator(
             new ClassMethodToCallbackExtractor(),
-            (new ParserFactory())->create(ParserFactory::PREFER_PHP7),
+            (new ParserFactory())->createForHostVersion(),
             new ClassDiscriminatorFromClassMetadata($classMetadataFactory),
             $allowReadOnlyTargetToPopulate
         ), __DIR__ . '/cache');